requirements.txt
streamlit>=1.28.0
chromadb>=0.4.15
googlemaps>=4.10.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0